Moved TrieBuilder to trie library from trie-builder.
