+++ libraries/trie/src/main/java/com/serwylo/lexica/trie/util/TrieBuilder.java
@@ -1,23 +1,30 @@
-package com.serwylo.lexica.trie;
+package com.serwylo.lexica.trie.util;
 
 import net.healeys.trie.Trie;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileReader;
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 
 public class TrieBuilder {
 
-	private final File usDictFile;
-	private final File ukDictFile;
+	private final InputStream usDictFile;
+	private final InputStream ukDictFile;
 	private final File outputFile;
 
-	public TrieBuilder(File usDictFile, File ukDictFile, File outputFile) {
-		this.usDictFile = usDictFile;
-		this.ukDictFile = ukDictFile;
+	public TrieBuilder(File usDictFile, File ukDictFile, File outputFile) throws IOException {
+		this(new FileInputStream(usDictFile), new FileInputStream(ukDictFile), outputFile);
+	}
+
+	public TrieBuilder(InputStream usDictFileStream, InputStream ukDictFileStream, File outputFile) {
+		this.usDictFile = usDictFileStream;
+		this.ukDictFile= ukDictFileStream;
 		this.outputFile = outputFile;
 	}
 
@@ -38,8 +45,8 @@ public class TrieBuilder {
 		}
 	}
 
-	private static void readFileIntoTrie(File dictFile, Trie trie, boolean usWord, boolean ukWord) throws IOException {
-		BufferedReader br = new BufferedReader(new FileReader(dictFile));
+	private static void readFileIntoTrie(InputStream dictFile, Trie trie, boolean usWord, boolean ukWord) throws IOException {
+		BufferedReader br = new BufferedReader(new InputStreamReader(dictFile));
 		String line;
 		while((line = br.readLine()) != null) {
 			trie.addWord(line, usWord, ukWord);
